[FIX]: throw an error if env variables are not declared

+++ src/services/deleteComment.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { CommentId } from "types";
 
-const URL = process.env.NEXT_PUBLIC_COMMENTS_API_URL ?? "";
+const URL = process.env.NEXT_PUBLIC_COMMENTS_API_URL;
 
 type Params = {
   commentId: CommentId;
@@ -9,5 +9,10 @@ type Params = {
 };
 
 export function deleteComment({ commentId, postId }: Params) {
+  if (!URL)
+    throw new Error(
+      "env variable 'NEXT_PUBLIC_COMMENTS_API_URL' is not declared"
+    );
+
   return axios.delete(`${URL}/${postId}`, { data: { commentId } });
 }
